Fix the test of whether disk names should be raw devices
Darwin and BSD platforms disk device names might have to be converted
to raw device names. There is a check for this in the code. But it is
missing a very important 'r'. This fix made it into the 0.9.7 release
but did not make the commit to this repo when this repo was given the
API changes for 0.9.7 because it was not clear whether the fix was
still relevant.