Fix crash in ifoRead_VTS_PTT_SRPT()
In ifoRead_VTS_PTT_SRPT(), there's a lot of places where the
function will exit on a 'goto fail' call if the IFO is invalid.

However, the check for the PGN value is only done later.

This fixes it so that it fails earlier -- when the PGN is
originally set, and doesn't go through the other checks as well, trying to set
invalid data that will be referenced later.

See VLC trac #10459