configure: fix build on NIOS II platform
NIOS II is a CPU architecture from Altera, which uses 'nios2' as the
architecture part of the tuple. Unfortunately, 'nios2' matches the
current '*os2*' test done by libdvdread's configure script to detect
the OS/2 operating system. This leads to build issues as the build
process of libdvdread then tries to use OS/2 specific compiler
options, that do not exist in the gcc used for Linux/NIOS2.

To fix this, this patch makes the test for OS/2 a little bit more
specific: in the case of the OS/2 operating system, the OS part of the
tuple contains just 'os2' (confirmed by looking at config.guess and
config.sub in the gnuconfig project). So using '*-os2-*' will properly
match the OS/2 operating system but not the NIOS II architecture.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>