configure.ac: Create an UNUSED #define for parameter passing
Some functions in libdvdread take parameters that are unused, e.g
file_read() in dvd_input.c. These functions can't be changed to remove
the unused parameters as they are public API.

GCC provides a __attribute__((unused)) annotation that can be used on
function parameters to inform the compiler of such and will squash
warnings of the following nature

    src/dvd_input.c: In function 'file_read':
    src/dvd_input.c:218:69: warning: unused parameter 'flags'
      [-Wunused-parameter]
     static int file_read(dvd_input_t dev, void *buffer, int blocks,
                          int flags)
                              ^

A check is added to configure.ac for the use of GCC and if found it
creates an UNUSED #define set to __attribute__((unused)) otherwise it
just sets it empty.

If compilers on other systems have a similar feature this check can be
adjusted to accommodate them.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>
Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>