src/dvd_input.c: Annotate unused function parameters
In libdvdread some functions take parameters that are unused. These
functions are for working without the libdvdcss library but have to have
the same prototype as the ones for working with it as these are
presented as single API.

As such we get the following warnings from GCC (4.8.3)

    src/dvd_input.c: In function 'file_error':
    src/dvd_input.c:186:37: warning: unused parameter 'dev'
      [-Wunused-parameter]
     static char *file_error(dvd_input_t dev)
                                         ^
    src/dvd_input.c: In function 'file_title':
    src/dvd_input.c:210:35: warning: unused parameter 'dev'
      [-Wunused-parameter]
     static int file_title(dvd_input_t dev, int block)
                                       ^
    src/dvd_input.c:210:44: warning: unused parameter 'block'
      [-Wunused-parameter]
     static int file_title(dvd_input_t dev, int block)
                                                ^
    src/dvd_input.c: In function 'file_read':
    src/dvd_input.c:218:69: warning: unused parameter 'flags'
      [-Wunused-parameter]
     static int file_read(dvd_input_t dev, void *buffer, int blocks,
                          int flags)
                              ^
The previous commit added a #define set to __attribute__((unused)) when
GCC is being used (empty otherwise). This can be used to inform the
compiler about such unused parameters and gets rid of those warnings.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>
Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>